requirements.txt
discord~=1.0.1
python-dotenv~=0.15.0
pyboy~=1.3.0
pip~=20.3.3
setuptools~=51.1.2
Cython~=0.29.21